pipecat/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pipecat-thymia-sentinel"
version = "1.0.0"
description = "Pipecat voice AI agent with Thymia Sentinel biomarker monitoring"
requires-python = ">=3.10"

dependencies = [
    "pipecat-ai[daily,openai,deepgram,cartesia,silero]>=0.0.50",
    "loguru>=0.7.3",
    "pydantic>=2.12.2",
    "python-dotenv",
    "websockets>=15.0.1",
    "fastapi",
    "uvicorn",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "A", "C4", "UP", "SIM", "RUF"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"